Change  model ID type to uint64

+++ model/model.go
@@ -8,7 +8,7 @@ import (
 
 //Model - used for embedding common database fields
 type Model struct {
-	ID        uint       `json:"id" gorm:"primary_key"`
+	ID        uint64     `json:"id" gorm:"primary_key"`
 	DeletedAt *time.Time `json:"-" sql:"index"`
 }
